+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.x branch
 ## 1.0 branch
 ### 1.0.6
+* added `file_exists_or_fail()`, `is_readable_or_fail()`, `is_writable_or_fail()`
+    functions;
 * added `assertIsArrayNotEmpty()` assertion method.
 
 ### 1.0.5